# api/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.ProjectMembership:
    properties:
      email:
        type: string
      id:
        type: string
      profile_id:
        type: string
      role:
        type: string
      role_id:
        type: string
      username:
        type: string
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.Submittal:
    properties:
      alert_config_id:
        type: string
      alert_config_name:
        type: string
      alert_type_id:
        type: string
      alert_type_name:
        type: string
      completion_date:
        type: string
      create_date:
        type: string
      due_date:
        type: string
      id:
        type: string
      marked_as_missing:
        type: boolean
      project_id:
        type: string
      submittal_status_id:
        type: string
      submittal_status_name:
        type: string
      warning_sent:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Monitoring Instrumentation Data Acquisition Systems (MIDAS) Web API
  license:
    name: MIT
    url: https://github.com/USACE/instrumentation-api/blob/555ea51191ff1245fe5910a295862be7514aaec6/LICENSE.md
  title: MIDAS Web API
  version: "2.0"
paths:
  /alert_configs/{alert_config_id}/submittals:
    get:
      parameters:
      - description: alert config uuid
        format: uuid
        in: path
        name: alert_config_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Submittal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists all submittals for an instrument
      tags:
      - submittal
  /alert_configs/{alert_config_id}/submittals/verify_missing:
    put:
      parameters:
      - description: alert config uuid
        format: uuid
        in: path
        name: alert_config_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: verifies all current submittals for the alert config are "missing"
        and will not be completed
      tags:
      - submittal
...
  /instruments/{instrument_id}/submittals:
    get:
      parameters:
      - description: instrument uuid
        format: uuid
        in: path
        name: instrument_id
        required: true
        type: string
      - description: filter by missing projects only
        in: query
        name: missing
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Submittal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists all submittals for an instrument
      tags:
      - submittal
...
  /projects/{project_id}/members:
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.ProjectMembership'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists project members and their role information
      tags:
      - project-role
  /projects/{project_id}/members/{profile_id}/roles/{role_id}:
    delete:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: profile uuid
        format: uuid
        in: path
        name: profile_id
        required: true
        type: string
      - description: role uuid
        format: uuid
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: removes project members and their role information
      tags:
      - project-role
    post:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: profile uuid
        format: uuid
        in: path
        name: profile_id
        required: true
        type: string
      - description: role uuid
        format: uuid
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.ProjectMembership'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: adds project members and their role information
      tags:
      - project-role
...
  /projects/{project_id}/submittals:
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: filter by missing projects only
        in: query
        name: missing
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Submittal'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists all submittals for a project
      tags:
      - submittal
...
  /submittals/{submittal_id}/verify_missing:
    put:
      parameters:
      - description: submittal uuid
        format: uuid
        in: path
        name: submittal_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: verifies the specified submittal is "missing" and will not be completed
      tags:
      - submittal
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"